Layout_TestCase/http/tests/xmlhttprequest/origin-whitelisting-ip-addresses-with-subdomains-expected.txt

CONSOLE MESSAGE: XMLHttpRequest cannot load http://http://172.16.7.123/test/jetpacktest/Layout_TestCase/http/tests/xmlhttprequest/resources/get.txt. Origin http://172.16.7.123/test/jetpacktest/Layout_TestCase/http/tests is not allowed by Access-Control-Allow-Origin.
CONSOLE MESSAGE: XMLHttpRequest cannot load http://http://172.16.7.123/test/jetpacktest/Layout_TestCase/http/tests/xmlhttprequest/resources/get.txt. Origin http://172.16.7.123/test/jetpacktest/Layout_TestCase/http/tests is not allowed by Access-Control-Allow-Origin.
Specifying that an IP address should match subdomains doesn't make sense. This test verifies that it doesn't do anything.

Testing: http://http://172.16.7.123/test/jetpacktest/Layout_TestCase/http/tests/xmlhttprequest/resources/get.txt (sync)
FAIL: Error: NETWORK_ERR: XMLHttpRequest Exception 101
Testing: http://http://172.16.7.123/test/jetpacktest/Layout_TestCase/http/tests/xmlhttprequest/resources/get.txt (async)
FAIL: 0
